Fix refresh rate and ball position in screen

+++ app/src/main/java/com/example/pong/model/Board.kt
@@ -1,20 +1,23 @@
 package com.example.pong.model
 
-import java.sql.Timestamp
-
 class Board(
-    private val width: Int,
-    private val height: Int,
+    private val width: Float,
+    private val height: Float,
     private val ball: Ball,
     private val brick: Brick,
 ) {
-    fun doStep(){
+    fun doStep() {
         ball.doStep()
         ball.checkCollision(width, height)
         brick.checkCollision(ball)
     }
 
-    fun applyAcceleration(xAcceleration: Float, yAcceleration: Float, zAcceleration: Float, timestamp: Long){
+    fun applyAcceleration(
+        xAcceleration: Float,
+        yAcceleration: Float,
+        zAcceleration: Float,
+        timestamp: Long
+    ) {
         brick.applyAcceleration(xAcceleration, yAcceleration, zAcceleration, width, timestamp)
     }
 }
